backend/requirements.txt
# Core Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23

# AI/ML & File Handling
openai==1.2.0
python-multipart    # FastAPI dependency for file uploads
python-dotenv      # To handle environment variables (API keys)
gensim==4.3.2
nltk==3.8.1
certifi==2023.7.22
numpy==1.24.3
scipy==1.10.1

# File Processing
python-docx==0.8.11 
PyPDF2==3.0.1       # for PDF support